+++ src/features/recipe/api/__tests__/apiSlice.test.ts
@@ -1,7 +1,7 @@
 import {describe, expect, it} from "vitest";
-import {useGetRecipesQuery} from "../apiSlice";
+import {useGetRecipesQuery} from "../recipesApi";
 import { act, renderHook } from '@testing-library/react';
-import {Recipe} from "../../features/recipe/state/recipesSlice";
+import {Recipe} from "../../state/editRecipeSlice";
 
 describe('test get requests for recipes', () => {
     it('request is succeuss', () => {
