+++ src/features/recipe/components/FormulaTable/Controls.tsx
@@ -1,6 +1,6 @@
 import {useAppDispatch, useAppSelector} from "../../../../store/hooks";
-import {addIngredient, } from "../../state/recipesSlice";
-import { selectFormula} from "../../state/recipesSelectors";
+import {addIngredient, } from "../../state/editRecipeSlice";
+import { selectFormula} from "../../state/editRecipeSelectors";
 import { v4 as uuidv4 } from 'uuid';
 import {useContext} from "react";
 import {RecipeContext} from "../../providers/RecipeProvider";
