+++ src/features/recipe/components/Yields/Yields.tsx
@@ -1,5 +1,5 @@
 import { useAppSelector } from "../../../../store/hooks";
-import { selectYields } from "../../state/recipesSelectors";
+import { selectYields } from "../../state/editRecipeSelectors";
 import { useContext } from "react";
 import { RecipeContext } from "../../providers/RecipeProvider";
 import Table from "../Table";
