+++ src/features/recipe/state/editRecipeSelectors.ts
@@ -1,80 +1,44 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { createStructuredSelector } from 'reselect'
-import {Formula, Ingredient, Preferment, Recipe} from "./recipesSlice";
 import { RootState } from "../../../store/store";
-import {Selector} from "react-redux/es/exports";
+import {Process, Yields, Formula, Ingredient, Preferment, Recipe} from "../types";
 
-export const selectRecipes = (state: RootState): Recipe[] => state.recipes;
+export const selectRecipe = (state: RootState): Recipe | undefined  => state.editRecipe;
+const selectYields = (state: RootState): Yields | undefined => state.editRecipe.yields
+const selectUnitWeight = (state: RootState): number | undefined => state.editRecipe.yields.unitWeight;
+const selectUnitQuantity = (state: RootState): number | undefined => state.editRecipe.yields.unitQuantity;
+const selectWasteFactor = (state: RootState): number | undefined => state.editRecipe.yields.wasteFactor;
+export const selectProcess = (state: RootState): Process | undefined => state.editRecipe.process;
+export const selectFormula = (state: RootState) => state.editRecipe.formula;
+const selectIngredients = (state: RootState) => state.editRecipe.formula.ingredients;
+const selectFlours = (state: RootState) => state.editRecipe.formula.flours;
 
-export const selectRecipe = (state: RootState, id: string): Recipe | undefined  => state.recipes.find(recipe => recipe!.id === id);
-
-export const selectFormula = createSelector(
-    selectRecipe,
-    (recipe) => recipe!.formula
-);
-
-export const selectIngredients = createSelector(
-    selectFormula,
-    (formula) => formula.ingredients
-);
-
-export const selectFlours = createSelector(
-    selectFormula,
-    (formula) => formula.flours
-);
-
-export const selectPreferments = createSelector(
-    selectRecipe,
-    (recipe) => recipe!.preferments
-)
 
 export const selectPreferment = (state: RootState, recipeId: string, prefermentId: string,): Preferment | undefined  =>
-    state.recipes.find(recipe => recipe!.id === recipeId)!
-        .preferments?.find(preferment => preferment!.id === prefermentId)
+    state.editRecipe.preferments?.find((preferment: Preferment) => preferment!.id === prefermentId);
 
 export const selectPrefermentTotalRatio = createSelector(
     selectPreferment,
     preferment =>
-        preferment!.formula.ingredients.reduce((totalRatio: number, ingredient: Ingredient) => totalRatio + ingredient.ratio, 1)
+        preferment!.formula.ingredients.reduce((totalRatio: number, ingredient: Ingredient) =>
+            totalRatio + ingredient.ratio, 1)
 );
 
-export const selectYields = createSelector(
-    selectRecipe,
-    recipe => recipe!.yields
-);
-
-export const selectProcess = createSelector(
-    selectRecipe,
-    recipe => recipe!.process
-)
-
-export const selectUnitWeight = createSelector(
-    selectYields,
-    yields => yields!.unitWeight
-);
-
-const selectUnitQuantity = createSelector(
-    selectYields,
-    yields => yields!.unitQuantity
-);
-
-const selectWasteFactor = createSelector(
-    selectYields,
-    yields => yields!.wasteFactor
-);
 
 const selectTotalRatio = createSelector(
-    selectIngredients, ingredients =>
-        ingredients.reduce((totalRatio: number, ingredient: Ingredient) => totalRatio + ingredient.ratio, 1)
+    selectIngredients,
+        ingredients =>
+        ingredients.reduce((totalRatio: number, ingredient: Ingredient) =>
+            totalRatio + ingredient.ratio, 1)
 );
 
 export const selectTotalWeight = createSelector(
     selectUnitWeight,
     selectUnitQuantity,
     selectWasteFactor,
-    ( unitWeight, unitQuantity, wasteFactor ) =>
-        Math.round(unitWeight * unitQuantity * (1 + wasteFactor))
-)
+    (unitWeight, unitQuantity, wasteFactor) =>
+        Math.round(unitWeight! * unitQuantity! * (1 + wasteFactor!))
+);
 
 export const selectTotals = createStructuredSelector({
     totalRatio: selectTotalRatio,
