+++ src/store/__tests__/recipeSlice.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { initialState, updateFlourRatio } from "../../features/recipe/state/recipesSlice";
+import { initialState, updateFlourRatio } from "../../features/recipe/state/editRecipeSlice";
 import {store} from "../store";
 
 describe('update ingredient ratios', () => {
