+++ src/store/store.ts
@@ -1,14 +1,14 @@
 import { configureStore } from '@reduxjs/toolkit';
-import recipesReducer from '../features/recipe/state/recipesSlice';
-import {apiSlice} from "../api/apiSlice";
+import editRecipeReducer from '../features/recipe/state/editRecipeSlice';
+import {recipesApi} from "../features/recipe/api/recipesApi";
 
 export const store = configureStore({
     reducer: {
-        recipe: recipesReducer,
-        [apiSlice.reducerPath]: apiSlice.reducer
+        editRecipe: editRecipeReducer,
+        [recipesApi.reducerPath]: recipesApi.reducer
     },
     middleware: (getDefaultMiddleware) => {
-        return getDefaultMiddleware().concat(apiSlice.middleware)
+        return getDefaultMiddleware().concat(recipesApi.middleware)
     }
 })
